refactor: move the progress handling to dasboard services.
