+++ backend/src/modules/livequizzes/controllers/PollRoomController.ts
@@ -356,10 +356,5 @@ export class PollRoomController {
   async getUserAchievements(@Param('userId') userId: string) {
     return await this.pollService.getUserAchievements(userId);
   }
-
-    @Get('/achievement/:userId/progress')
-  async getUserAchievementProgress(@Param('userId') userId: string) {
-    return await this.pollService.getUserAchievementProgress(userId);
-  }
   
 }
